refactor(msfem): compute the AMS edge extensions separately
As it was implemented, for large problems, a dense matrix would be alocated for `Phi_e`, leading to a memory issue. Now, the basis functions on the edge are computed for each coarse node and filtered in place. Furthermore, to improve the performance, a factorization of `A_ee` is computed.